Rename: 5팀 컴포넌트이름 시작 Assignment 붙임
