+++ src/app/assignment/(components)/AssignmentProfileImage.tsx
